Move admin routes into blueprint
and update templates

Routes:
* /reports
* /adminlog
