+++ nyaa/views/__init__.py
@@ -1,7 +1,9 @@
 from nyaa.views import (
     account,
+    admin,
     site,
 )
 
 account_bp = account.bp
+admin_bp = admin.bp
 site_bp = site.bp
